Separate euality comparer tests file
